const-maccro has to be size specified not usize

+++ programs/nft_marketplace/src/state/marketplace.rs
@@ -15,5 +15,5 @@ pub struct Marketplace {
 }
 
 impl Space for Marketplace {
-    const INIT_SPACE: usize = ANCHOR_DISCRIMINATOR + 32 + 2 + 1 + 1 + 1 + (4 + 32);
+    const INIT_SPACE: usize = (ANCHOR_DISCRIMINATOR as usize) + 32 + 2 + 1 + 1 + 1 + (4 + 32);
 }
